Refactor based on new knowledge
Also make the CLI a bit easier to use.

The properties can be optionally prefixed with 0xe0 -- and the size
changes cause of it, and is included in the 0x8X byte. The byte after
that is simply whether to set a value, or load it into the report buffer
on the next GET_REPORT.